Move MidtransSdk to top of package and test implementation in app-MainActivity
